+++ HISTORY.md
@@ -1,5 +1,27 @@
 # History
 
+
+## 2015-05-09, version 0.8.1
+
+- Implemented option `typed.ignore` to ignore/filter signatures of a typed
+  function.
+
+
+## 2015-05-09, version 0.8.0
+
+- Implemented function `create` to create a new instance of typed-function.
+- Implemented a utility function `convert(value, type)` (#1).
+- Implemented a simple `typed.find` function to find the implementation of a
+  specific function signature.
+- Extended the error messages to denote the function name, like `"Too many 
+  arguments in function foo (...)"`.
+
+
+## 2015-04-17, version 0.7.0
+
+- Performance improvements.
+
+
 ## 2015-03-08, version 0.6.3
 
 - Fixed generated internal Signature and Param objects not being cleaned up